rename some existing mixins
